# docs/contracts/openapi-v1.yaml
openapi: 3.1.0
...
info:
  title: AI Body Companion API
  version: 1.1.0-draft
  description: |
    面向 iOS 的身体信号记录与恢复决策助手 API。

    Agent 输出只是不可信候选数据。确定性安全规则、身份认证、对象级授权、
    用户确认、幂等、事务写入和审计均由服务端负责。
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.bodycompanion.example
    description: Production placeholder
  - url: https://staging-api.bodycompanion.example
    description: Staging placeholder
security:
  - BearerAuth: []
...
paths:
...
  /v1/agent/sessions:
    post:
      tags: [Assessment Sessions]
      operationId: startAssessment
      summary: 创建身体信号 AssessmentSession
      description: >-
        服务端必须验证当前用户存在仍有效、由用户确认的 age_eligibility=adult_18_plus，
        并验证核心记录/云端处理所需同意；不得信任客户端 onboarding 状态。
        未确认、已撤回或明确不符合年龄范围时不得创建 Session 或持久化健康输入。
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAssessmentRequest'
      responses:
        '201':
          description: Session 已创建
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/agent/sessions/{session_id}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [Assessment Sessions]
      operationId: getAssessment
      summary: 获取或恢复 AssessmentSession
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
      responses:
        '200':
          description: 当前 Session 状态；不包含 Provider 原始消息或隐藏推理
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Assessment Sessions]
      operationId: discardAssessmentDraft
      summary: 丢弃当前 Session 的未确认草稿并使会话失效
      description: >-
        只允许处理尚未持久化为正式 Event 的草稿；已确认历史不受影响，操作写入最小审计记录。
        若 Session 正在 awaiting_approval，服务端必须在同一事务将该 Session 尚未执行的 ApprovalIntent
        标为 invalidated；旧 approval_id 后续决定必须无副作用地拒绝。
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequiredIfMatch'
      responses:
        '204':
          description: 未确认草稿已丢弃、相关未执行审批已失效；重复调用保持成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/agent/sessions/{session_id}/turns:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [Assessment Sessions]
      operationId: submitAssessmentAnswer
      summary: 提交一轮用户输入并异步运行受控 Agent
      description: |
        Turn 只能生成候选结构或下一步动作，不能直接创建正式 BodySignalEvent。
        同一 Session 同时只允许一个运行中的 Turn。
        每次提交都重新验证当前 age_eligibility 和所需 ConsentScope；撤回后不得继续处理旧 Session。
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentTurnRequest'
      responses:
        '202':
          description: Turn 已排队或正在运行
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentTurn'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/UpstreamTimeout'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/agent/sessions/{session_id}/confirmations:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [Assessment Sessions]
      operationId: requestDraftConfirmation
      summary: 提交用户复核后的候选事实并创建审批意图
      description: |
        本接口不会直接写入正式身体事件。生产实现必须先复用已验证的 P2B typed ConfirmationIntent
        application boundary，重新执行 Schema、本体、业务和安全校验，再在同一事务创建 ApprovalIntent、
        application-origin approval_required Turn，
        并把 Session 推进 awaiting_approval、更新 latest_turn_id。
        服务端必须再次验证当前 age_eligibility、同意和安全状态。
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConfirmationIntentRequest'
      responses:
        '201':
          description: 审批意图已创建，尚未执行正式写入
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmationIntentResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/approvals/{approval_id}:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    get:
      tags: [Approvals]
      operationId: getApproval
      summary: 恢复当前用户的审批意图
      description: |
        仅返回当前用户拥有且仍可见的审批意图，用于中断、跨设备和 Location 恢复。
        pending 可继续批准/拒绝；终态只能查看，不能再次决定。来源 TTL 内可返回完整终态结果；
        来源清理后只返回 redacted ApprovalIntent 墓碑且 decision_result=null，不含 display_summary/Session/Turn 正文。
        不存在与无权访问统一返回 404。
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
      responses:
        '200':
          description: 审批意图当前快照
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRecoveryResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/body-signal-events/{event_id}:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Body Signal Events]
      operationId: getBodySignalEvent
      summary: 获取已确认的身体信号事件
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ClientVersion'
      responses:
        '200':
          description: 事件快照；AI 解释与用户事实保持分层
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BodySignalEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token；资源所有者由服务端认证上下文决定。
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: 客户端请求 UUID；服务端会返回最终 request ID。
      schema:
        type: string
        format: uuid
    ClientVersion:
      name: X-Client-Version
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 40
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        maxLength: 64
        example: zh-CN
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        同一用户动作重试必须复用。不含健康正文的 key/作用域/请求哈希/执行状态/不透明结果引用至少保留 7 天；
        可重放的响应正文保留期不得长于其来源 Session/草稿/同意 TTL，清理后同 key 只返回稳定墓碑/410，不还原旧正文。
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[!-~]{8,128}$'
...
  headers:
    RequestId:
      description: 服务端最终 request ID。
      schema:
        type: string
        format: uuid
    ETag:
      description: 资源 revision 的强 ETag。
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    Location:
      description: 新建资源或异步资源的相对/绝对位置。
      schema:
        type: string
        format: uri-reference
...
    IdempotencyReplayed:
      description: true 表示返回先前已提交动作的结果。
      schema:
        type: boolean
  schemas:
...
    AgentTurn:
      $ref: ./agent-turn.schema.json
...
    StartAssessmentRequest:
      type: object
      additionalProperties: false
      required: [purpose, locale, timezone, client_context]
      properties:
        purpose:
          type: string
          const: body_signal_assessment
        episode_id:
          type: [string, 'null']
          format: uuid
        locale:
          type: string
          pattern: '^[a-z]{2,3}(?:-[A-Z]{2})?$'
        timezone:
          type: string
          minLength: 1
          maxLength: 64
        client_context:
          $ref: '#/components/schemas/ClientContext'
...
  responses:
    Unauthorized:
      description: 未认证或 token 无效
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 缺少近期认证、有效同意或当前动作权限
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在或当前用户无权获知其存在
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...